+++ src/components/base/Base.js
@@ -2687,6 +2687,14 @@ export default class BaseComponent extends Component {
       attributes.tabindex = this.component.tabindex;
     }
 
+    if (this.component.htmlAttributes) {
+      this.component.htmlAttributes.forEach(attribute => {
+        if (!attributes.hasOwnProperty(attribute.name)) {
+          attributes[attribute.name] = attribute.value;
+        }
+      });
+    }
+
     return {
       type: 'input',
       component: this.component,